# Wavelet/docs/swagger.yaml
basePath: /
definitions:
  apply_log.CleanupInput:
    properties:
      delete_all:
        type: boolean
      retention_days:
        type: integer
    type: object
  apply_log.CleanupResult:
    properties:
      cutoff:
        type: string
      delete_all:
        type: boolean
      deleted_count:
        type: integer
      retention_days:
        type: integer
    type: object
  apply_log.ListResult:
    properties:
      current:
        type: integer
      rows:
        items:
          $ref: '#/definitions/model.OpenFlareApplyLog'
        type: array
      total:
        type: integer
      totalPage:
        type: integer
    type: object
...
  config_version.CleanupInput:
    properties:
      keep_count:
        type: integer
    type: object
  config_version.CleanupResult:
    properties:
      deleted_count:
        type: integer
      message:
        type: string
    type: object
  config_version.ConfigDiffResult:
    properties:
      active_version:
        type: string
      active_website_count:
        type: integer
      added_domains:
        items:
          type: string
        type: array
      added_sites:
        items:
          type: string
        type: array
      changed_option_details:
        items:
          $ref: '#/definitions/config_version.ConfigOptionDiffItem'
        type: array
      changed_option_keys:
        items:
          type: string
        type: array
      current_website_count:
        type: integer
      main_config_changed:
        type: boolean
      modified_domains:
        items:
          type: string
        type: array
      modified_sites:
        items:
          type: string
        type: array
      removed_domains:
        items:
          type: string
        type: array
      removed_sites:
        items:
          type: string
        type: array
      waf_config_changed:
        type: boolean
    type: object
  config_version.ConfigOptionDiffItem:
    properties:
      current_value:
        type: string
      key:
        type: string
      previous_value:
        type: string
    type: object
  config_version.ConfigPreviewResult:
    properties:
      checksum:
        type: string
      main_config:
        type: string
      rendered_config:
        type: string
      route_config:
        type: string
      route_count:
        type: integer
      snapshot_json:
        type: string
      support_files:
        items:
          $ref: '#/definitions/config_version.SupportFile'
        type: array
      website_count:
        type: integer
    type: object
  config_version.SupportFile:
    properties:
      content:
        type: string
      path:
        type: string
    type: object
  dashboard.Capacity:
    properties:
      average_cpu_usage_percent:
        type: number
      average_memory_usage_percent:
        type: number
      high_cpu_nodes:
        type: integer
      high_memory_nodes:
        type: integer
      high_storage_nodes:
        type: integer
    type: object
  dashboard.OverviewPayload:
    properties:
      capacity:
        $ref: '#/definitions/dashboard.Capacity'
      distributions:
        $ref: '#/definitions/dashboard.distributionsPayload'
      generated_at: {}
      nodes:
        items:
          items: {}
          type: array
        type: array
      summary:
        $ref: '#/definitions/dashboard.Summary'
      traffic:
        $ref: '#/definitions/dashboard.Traffic'
      trends:
        $ref: '#/definitions/dashboard.trendsPayload'
    type: object
  dashboard.Summary:
    properties:
      offline_nodes:
        type: integer
      online_nodes:
        type: integer
      pending_nodes:
        type: integer
      total_nodes:
        type: integer
      unhealthy_nodes:
        type: integer
    type: object
  dashboard.Traffic:
    properties:
      error_count:
        type: integer
      estimated_qps:
        type: number
      reported_nodes:
        type: integer
      request_count:
        type: integer
      unique_visitors:
        type: integer
    type: object
  dashboard.distributionsPayload:
    properties:
      source_countries:
        items:
          items: {}
          type: array
        type: array
      status_codes:
        items:
          items: {}
          type: array
        type: array
      top_domains:
        items:
          items: {}
          type: array
        type: array
    type: object
  dashboard.trendsPayload:
    properties:
      capacity_24h:
        items:
          items: {}
          type: array
        type: array
      disk_io_24h:
        items:
          items: {}
          type: array
        type: array
      network_24h:
        items:
          items: {}
          type: array
        type: array
      traffic_24h:
        items:
          items: {}
          type: array
        type: array
    type: object
...
  model.ConfigVersion:
    properties:
      checksum:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      main_config:
        type: string
      rendered_config:
        type: string
      snapshot_json:
        type: string
      support_files_json:
        type: string
      version:
        type: string
    type: object
  model.ConfigVersionSummary:
    properties:
      checksum:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      version:
        type: string
    type: object
...
  model.OpenFlareApplyLog:
    properties:
      checksum:
        type: string
      created_at:
        type: string
      id:
        type: integer
      main_config_checksum:
        type: string
      message:
        type: string
      node_id:
        type: string
      result:
        type: string
      route_config_checksum:
        type: string
      support_file_count:
        type: integer
      version:
        type: string
    type: object
...
  response.Any:
    properties:
      data: {}
      error_msg:
        example: ""
        type: string
    type: object
...
info:
  contact:
    name: OpenFlare
    url: https://github.com/Rain-kl/OpenFlare
  description: OpenFlare 平台后端 API，提供用户认证、系统配置、任务调度与边缘节点管理能力。
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: OpenFlare API
  version: 1.0.0
paths:
...
  /api/v1/openflare/apply-logs:
    get:
      description: 分页返回节点配置下发记录，支持按节点 ID 筛选，需要管理员权限
      parameters:
      - description: 节点 ID 筛选
        in: query
        name: node_id
        type: string
      - description: 页码
        in: query
        name: pageNo
        type: integer
      - description: 页码（别名）
        in: query
        name: page_no
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 每页数量（别名）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 下发日志列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/apply_log.ListResult'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 无权限或不存在
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 获取配置下发日志
      tags:
      - openflare-apply-log
  /api/v1/openflare/apply-logs/cleanup:
    post:
      consumes:
      - application/json
      description: 按保留天数清理历史下发记录，或删除全部记录，需要管理员权限
      parameters:
      - description: 清理参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apply_log.CleanupInput'
      produces:
      - application/json
      responses:
        "200":
          description: 清理结果
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/apply_log.CleanupResult'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 无权限或不存在
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 清理配置下发日志
      tags:
      - openflare-apply-log
  /api/v1/openflare/config-versions:
    get:
      description: 返回所有已发布的 OpenResty 配置版本摘要，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: 配置版本列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ConfigVersionSummary'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 无权限或不存在
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 获取配置版本列表
      tags:
      - openflare-config-version
  /api/v1/openflare/config-versions/{id}:
    get:
      description: 返回指定配置版本的完整快照与渲染内容，需要管理员权限
      parameters:
      - description: 配置版本 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 配置版本详情
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/model.ConfigVersion'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 无权限或版本不存在
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 获取配置版本详情
      tags:
      - openflare-config-version
  /api/v1/openflare/config-versions/{id}/activate:
    post:
      description: 将指定历史版本设为当前活跃配置，需要管理员权限
      parameters:
      - description: 配置版本 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 激活成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/model.ConfigVersion'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 无权限或版本不存在
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 激活配置版本
      tags:
      - openflare-config-version
  /api/v1/openflare/config-versions/active:
    get:
      description: 返回当前正在使用的配置版本，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: 活跃配置版本
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/model.ConfigVersion'
              type: object
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 无权限、不存在或无活跃版本
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 获取当前活跃配置版本
      tags:
      - openflare-config-version
  /api/v1/openflare/config-versions/cleanup:
    post:
      consumes:
      - application/json
      description: 删除超出保留数量的非活跃配置版本，需要管理员权限
      parameters:
      - description: 清理参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/config_version.CleanupInput'
      produces:
      - application/json
      responses:
        "200":
          description: 清理结果
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/config_version.CleanupResult'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 无权限或不存在
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 清理历史配置版本
      tags:
      - openflare-config-version
  /api/v1/openflare/config-versions/diff:
    get:
      description: 对比当前草稿配置与活跃版本之间的差异，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: 配置差异
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/config_version.ConfigDiffResult'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 无权限或不存在
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 对比草稿与活跃配置
      tags:
      - openflare-config-version
  /api/v1/openflare/config-versions/preview:
    get:
      description: 渲染并返回当前草稿配置的预览结果，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: 配置预览
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/config_version.ConfigPreviewResult'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 无权限或不存在
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 预览当前草稿配置
      tags:
      - openflare-config-version
  /api/v1/openflare/config-versions/publish:
    post:
      description: 将当前草稿配置发布为新版本，需要管理员权限
      parameters:
      - description: 是否强制发布
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 发布成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/model.ConfigVersion'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 无权限或不存在
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 发布配置版本
      tags:
      - openflare-config-version
  /api/v1/openflare/dashboard/overview:
    get:
      description: 聚合节点与可观测性数据，返回 OpenFlare 控制台仪表盘概览，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: 仪表盘概览
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/dashboard.OverviewPayload'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 获取仪表盘概览
      tags:
      - openflare-dashboard
...
securityDefinitions:
  SessionCookie:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"